LPS-182364 Regen services

+++ modules/apps/portal/portal-uad/src/main/java/com/liferay/portal/uad/anonymizer/BaseRepositoryEntryUADAnonymizer.java
@@ -57,8 +57,12 @@ public abstract class BaseRepositoryEntryUADAnonymizer
 	}
 
 	@Override
-	public void delete(RepositoryEntry repositoryEntry) throws PortalException {
-		repositoryEntryLocalService.deleteRepositoryEntry(repositoryEntry);
+	public void delete(RepositoryEntry repositoryEntry, long userId)
+		throws PortalException {
+
+		if (repositoryEntry.getUserId() == userId) {
+			repositoryEntryLocalService.deleteRepositoryEntry(repositoryEntry);
+		}
 	}
 
 	@Override
